chore: Modal 내부 추상화 (ModalBody, ModalFooter)
