+++ packages/ui/src/Modal/Modal.tsx
@@ -2,18 +2,11 @@
 
 import { Dialog, DialogContainer, DialogContent, Portal } from "@ark-ui/react";
 
-import { typeToIcon } from "./Modal.constants";
 import * as Styled from "./Modal.styles";
 import type { ModalProps } from "./Modal.types";
+import { ModalBody, ModalFooter } from "./components";
 
 export const Modal = ({ isOpen, props, onClose }: ModalProps) => {
-  const { title, type, onConfirm } = props;
-
-  const onConfirmAndClose = () => {
-    onConfirm();
-    onClose();
-  };
-
   return (
     <Dialog open={isOpen} onClose={onClose}>
       <Portal>
@@ -21,23 +14,8 @@ export const Modal = ({ isOpen, props, onClose }: ModalProps) => {
           <Styled.Background>
             <DialogContent>
               <Styled.ModalWrapper>
-                <Styled.Body>
-                  <span
-                    className="material-symbols-outlined"
-                    style={{ fontSize: "48px", color: typeToIcon[type].color }}
-                  >
-                    {typeToIcon[type].icon}
-                  </span>
-                  <Styled.Text>{title}</Styled.Text>
-                  {type === "warning" ? (
-                    <Styled.ButtonWrap>
-                      <Styled.Button onClick={onClose}>아니오</Styled.Button>
-                      <Styled.Button onClick={onConfirmAndClose}>예</Styled.Button>
-                    </Styled.ButtonWrap>
-                  ) : (
-                    <Styled.Button onClick={onConfirmAndClose}>확인</Styled.Button>
-                  )}
-                </Styled.Body>
+                <ModalBody props={props} />
+                <ModalFooter props={props} onClose={onClose} />
               </Styled.ModalWrapper>
             </DialogContent>
           </Styled.Background>
