+++ packages/ui/src/Modal/components/index.ts
@@ -1,0 +1,2 @@
+export { ModalBody } from "./ModalBody";
+export { ModalFooter } from "./ModalFooter";
